hv/mshv_vtl: Port and Support arm64 Architecture
Extend MSHV_VTL support to arm64 architecture, ensuring that
virtual trust levels are comprehensively supported across
hardware platforms.
