+++ drivers/hv/mshv_vtl.h
@@ -4,33 +4,7 @@
 
 #include <linux/mshv.h>
 #include <linux/types.h>
-#include <asm/fpu/types.h>
-
-struct mshv_vtl_cpu_context {
-	union {
-		struct {
-			u64 rax;
-			u64 rcx;
-			u64 rdx;
-			u64 rbx;
-			u64 cr2;
-			u64 rbp;
-			u64 rsi;
-			u64 rdi;
-			u64 r8;
-			u64 r9;
-			u64 r10;
-			u64 r11;
-			u64 r12;
-			u64 r13;
-			u64 r14;
-			u64 r15;
-		};
-		u64 gp_regs[16];
-	};
-
-	struct fxregs_state fx_state;
-};
+#include <asm/mshyperv.h>
 
 struct mshv_vtl_run {
 	u32 cancel;
@@ -41,7 +15,7 @@ struct mshv_vtl_run {
 	__u8 enter_mode;
 	char exit_message[MAX_RUN_MSG_SIZE];
 	union {
-		struct mshv_vtl_cpu_context cpu_context;
+		struct hv_vtl_cpu_context cpu_context;
 
 		/*
 		 * Reserving room for the cpu context to grow and be
